Conflicts:
	include/of2g.h
	src/main.c
	src/of2g.c

+++ include/netman.h
@@ -65,7 +65,7 @@ void netman_init(netman_t * netman);
 
 // Build an of2g DATA frame based on the given buffer of raw data, and
 // store it in netman->current_tx_data
-void netman_new_tx_bytes(netman_t * netman, unsigned char * buffer, uint8_t length);
+void netman_new_tx_bytes(netman_t * netman, unsigned char * buffer, size_t length);
 
 // Build an of2g ACK frame based on the given data frame
 void build_ack_frame (netman_t * netman, of2g_frame_t * data_frame);
